requirements.txt
wheel==0.42.0
setuptools==69.0.3
numpy==1.24.3
pandas==2.0.1
flask==3.0.0
--find-links https://download.pytorch.org/whl/torch_stable.html
torch==2.1.0+cpu
transformers==4.37.2
sentence-transformers==2.3.1
faiss-cpu==1.7.4
google-generativeai==0.3.2
python-dotenv==1.0.0
pymongo==4.6.1
certifi==2024.2.2